[Dev] Channel decomposition.
